Move the file that builds the TCF Experience contents into the new tcf folder.
