+++ src/fides/api/models/privacy_preference.py
@@ -34,7 +34,7 @@ from fides.api.models.privacy_request import (
     ProvidedIdentity,
 )
 from fides.api.models.sql_models import System  # type: ignore[attr-defined]
-from fides.api.util.tcf_util import (
+from fides.api.util.tcf.tcf_experience_contents import (
     ConsentRecordType,
     TCFComponentType,
     get_relevant_systems_for_tcf_attribute,
