+++ src/fides/api/util/tcf/experience_meta.py
@@ -9,7 +9,7 @@ from fides.api.schemas.base_class import FidesSchema
 from fides.api.schemas.privacy_experience import ExperienceMeta, TCMobileData
 from fides.api.util.tcf.tc_mobile_data import build_tc_data_for_mobile
 from fides.api.util.tcf.tc_model import TCModel, convert_tcf_contents_to_tc_model
-from fides.api.util.tcf_util import TCFExperienceContents
+from fides.api.util.tcf.tcf_experience_contents import TCFExperienceContents
 
 
 class TCFVersionHash(FidesSchema):
