+++ src/fides/api/util/tcf/tc_model.py
@@ -8,7 +8,7 @@ from pydantic import Field, NonNegativeInt, PositiveInt, root_validator, validat
 from fides.api.models.privacy_notice import UserConsentPreference
 from fides.api.schemas.base_class import FidesSchema
 from fides.api.schemas.tcf import TCFFeatureRecord, TCFPurposeRecord, TCFVendorRecord
-from fides.api.util.tcf_util import TCFExperienceContents, load_gvl
+from fides.api.util.tcf.tcf_experience_contents import TCFExperienceContents, load_gvl
 
 CMP_ID: int = 12  # TODO: hardcode our unique CMP ID after certification
 CMP_VERSION = 1
